+++ app/code/Kemana/Survey/Api/IspApiInterface.php
@@ -1,0 +1,11 @@
+<?php
+namespace Kemana\Survey\Api;
+
+
+interface IspApiInterface
+{
+    /**
+     * @return array|null
+     */
+    public function getIspData();
+}
