+++ app/code/Kemana/Survey/etc/csp_whitelist.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0"?>
+<csp_whitelist xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Csp:etc/csp_whitelist.xsd">
+    <policies>
+        <policy id="script-src">
+            <values>
+                <value id="devdocs-base" type="host">https://developer.adobe.com</value>
+                <value id="magento" type="host">https://magento.com</value>
+                <value id="ips" type="host">http://ip-api.com</value>
+            </values>
+        </policy>
+        <policy id="connect-src">
+            <values>
+                <value id="devdocs" type="host">https://developer.adobe.com</value>
+                <value id="ips" type="host">http://ip-api.com</value>
+            </values>
+        </policy>
+    </policies>
+</csp_whitelist>
